vehicle persistence
Vehicle persistence with inventory, damage, and unique variables, such as HG_Owner (vehicle garages) etc. including static weapons. The one thing that whooped my ass was the addaction for claiming the vehicle. I may need to rethink my process. I am trying to add a check. If there is an owner, the claim button code should not execute. Maybe I need rethink the entire process.

maybe create a button that says "check for owner" than after that check, it will create a claim button or if there is an owner, say sorry, this car is aleady owned...